scheduling/requirements.txt
adium-theme-ubuntu==0.3.4
backports.functools-lru-cache==1.4
backports.shutil-get-terminal-size==1.0.0
bleach==2.1.1
configparser==3.5.0
cycler==0.10.0
decorator==4.0.6
entrypoints==0.2.3
enum34==1.1.6
functools32==3.2.3.post2
html5lib==1.0b10
ipykernel==4.6.1
ipython==5.5.0
ipython-genutils==0.2.0
ipywidgets==7.0.3
Jinja2==2.8
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.1.0
jupyter-console==5.2.0
jupyter-core==4.4.0
MarkupSafe==0.23
matplotlib==2.1.0
mistune==0.8
nbconvert==5.3.1
nbformat==4.4.0
notebook==5.2.0
numpy==1.13.3
olefile==0.44
pandocfilters==1.4.2
pathlib2==2.3.0
pexpect==4.0.1
pickleshare==0.7.4
Pillow==4.3.0
prompt-toolkit==1.0.15
ptyprocess==0.5
pycurl==7.43.0
Pygments==2.2.0
pygobject==3.20.0
PyMySQL==0.7.2
pyparsing==2.2.0
python-dateutil==2.6.1
pytz==2017.3
pyzmq==15.2.0
qtconsole==4.3.1
scandir==1.6
simplegeneric==0.8.1
six==1.11.0
subprocess32==3.2.7
terminado==0.6
testpath==0.3.1
tornado==4.2.1
traitlets==4.3.2
unity-lens-photos==1.0
wcwidth==0.1.7
webencodings==0.5.1
widgetsnbextension==3.0.6
